De-noise forge factory
Compares configs when updating the forge factory to avoid recreating 
forge clients and causing unnecessary refresh commands.
